simplify datagen code
