+++ datagen/src/dataformat.rs
@@ -1,7 +1,5 @@
 mod binpack;
 mod board;
-mod policy;
 
 pub use binpack::Binpack;
 pub use board::CompressedChessBoard;
-pub use policy::PolicyData;
